Added support for directly-interpolated waveforms
Added support for VP_SHAPE='alt spline', which directly interpolates waveforms
for the vector-potential. This form is intended as a replacement for the more
restrictive VP_SHAPE='table', and should be particularly useful for calculating
RABBITT and streaking spectra.

Also added new config files for gfortran 12 and Intel OneAPI compilers, and
made some cosmetic changes to the makefiles.